+++ app/Filament/Resources/StockLevels/Schemas/StockLevelForm.php
@@ -7,11 +7,16 @@ namespace Modules\ERP\Filament\Resources\StockLevels\Schemas;
 use Filament\Forms\Components\Select;
 use Filament\Forms\Components\TextInput;
 use Filament\Schemas\Schema;
+use Modules\Core\Filament\Utils\HasForm;
 
 final class StockLevelForm
 {
+    use HasForm;
+
     public static function configure(Schema $schema): Schema
     {
+        self::configureForm($schema);
+
         return $schema
             ->components([
                 Select::make('company_id')