+++ app/Models/ReturnOrderLine.php
@@ -22,7 +22,6 @@ use Override;
  * @property numeric-string $quantity
  * @property numeric-string|null $unit_cost
  * @property numeric-string|null $unit_price
- *
  * @mixin \Eloquent
  * @mixin IdeHelperReturnOrderLine
  */